third_party/WebKit/LayoutTests/crypto/gc-2-expected.txt

Test that window.crypto.subtle wrapper preserves custom properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS [object Crypto] is defined.
PASS [object SubtleCrypto] is defined.
PASS anotherWindowCrypto.subtle.foo is "bar"
PASS successfullyParsed is true

TEST COMPLETE
